# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      replies:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      reply:
        $ref: '#/definitions/entity.Comment'
      reply_id:
        type: integer
      submission:
        $ref: '#/definitions/entity.Submission'
      submission_id:
        type: integer
      text:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.Contest:
    properties:
      created_at:
        type: string
      id:
        type: integer
      link:
        type: string
      link2:
        type: string
      link3:
        type: string
      name:
        type: string
      problem_count:
        type: integer
      problems:
        description: Relations
        items:
          $ref: '#/definitions/entity.Problem'
        type: array
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      type:
        type: string
      unrated:
        type: boolean
      updated_at:
        type: string
      votes:
        description: |-
          Ratings       []Rating       `json:"ratings,omitempty" gorm:"foreignKey:ContestID"`
          DivisionUsers []DivisionUser `json:"division_users,omitempty" gorm:"foreignKey:ContestID"`
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.Country:
    properties:
      created_at:
        type: string
      groups:
        items:
          $ref: '#/definitions/entity.Group'
        type: array
      id:
        type: integer
      name:
        type: string
      short_code:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.DailyProblem:
    properties:
      created_at:
        type: string
      for_date:
        type: string
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Exercise:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  entity.GoogleOAuth:
    properties:
      calendar_id:
        type: string
      created_at:
        type: string
      encrypted_token_string:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Group:
    properties:
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      google_oauths:
        items:
          $ref: '#/definitions/entity.GoogleOAuth'
        type: array
      group_sessions:
        items:
          $ref: '#/definitions/entity.GroupSession'
        type: array
      hoa_id:
        description: Head of Academy ID
        type: integer
      hoas:
        items:
          $ref: '#/definitions/entity.HOA'
        type: array
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/entity.Invite'
        type: array
      name:
        type: string
      short_name:
        type: string
      super_to_groups:
        items:
          $ref: '#/definitions/entity.SuperToGroup'
        type: array
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.GroupSession:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      session:
        $ref: '#/definitions/entity.Session'
      session_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.HOA:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Invite:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      key:
        description: Unique invite key
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      used:
        description: Whether the invite has been used
        type: boolean
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        description: User who created the invite
        type: integer
    type: object
  entity.Post:
    properties:
      body:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      post_to_tags:
        items:
          $ref: '#/definitions/entity.PostToTag'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
...
  entity.Problem:
    properties:
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      contest:
        $ref: '#/definitions/entity.Contest'
      contest_id:
        type: integer
      created_at:
        type: string
      daily_problems:
        items:
          $ref: '#/definitions/entity.DailyProblem'
        type: array
      difficulty:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      platform:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      submissions:
        description: Relations
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      tag:
        type: string
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.ProblemTrack:
    properties:
      created_at:
        type: string
      id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.RecentAction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      invites:
        items:
          $ref: '#/definitions/entity.Invite'
        type: array
      type:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
...
  entity.Submission:
    properties:
      code:
        type: string
      comments:
        description: Relations
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      in_contest:
        description: Whether it was solved in a contest
        type: integer
      language:
        type: string
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      time_spent:
        description: Time spent in seconds
        type: integer
      tries:
        description: Number of attempts
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
      verified:
        type: boolean
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.SuperGroup:
    properties:
      contests:
        items:
          $ref: '#/definitions/entity.Contest'
        type: array
      created_at:
        type: string
      daily_problems:
        items:
          $ref: '#/definitions/entity.DailyProblem'
        type: array
      id:
        type: integer
      name:
        type: string
      super_to_groups:
        items:
          $ref: '#/definitions/entity.SuperToGroup'
        type: array
      tracks:
        description: Relations
        items:
          $ref: '#/definitions/entity.Track'
        type: array
      updated_at:
        type: string
    type: object
  entity.SuperToGroup:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      id:
        type: integer
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
    type: object
  entity.Track:
    properties:
      active:
        type: boolean
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      created_at:
        type: string
      exercises:
        items:
          $ref: '#/definitions/entity.Exercise'
        type: array
      id:
        type: integer
      name:
        type: string
      problem_tracks:
        items:
          $ref: '#/definitions/entity.ProblemTrack'
        type: array
      problems:
        description: Relations
        items:
          $ref: '#/definitions/entity.Problem'
        type: array
      super_group:
        $ref: '#/definitions/entity.SuperGroup'
      super_group_id:
        type: integer
      updated_at:
        type: string
      votes:
        items:
          $ref: '#/definitions/entity.Vote'
        type: array
    type: object
  entity.User:
    properties:
      api_tokens:
        items:
          $ref: '#/definitions/entity.APIToken'
        type: array
      attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      birthday:
        description: Personal Information
        type: string
      code_of_conduct:
        type: string
      codeforces:
        type: string
      comments:
        items:
          $ref: '#/definitions/entity.Comment'
        type: array
      config:
        type: string
      country:
        $ref: '#/definitions/entity.Country'
      country_id:
        type: integer
      created_at:
        description: Timestamps
        type: string
      cv:
        description: Professional Details
        type: string
      department:
        type: string
      email:
        type: string
      expected_graduation_date:
        type: string
      gender:
        type: string
      github:
        type: string
      group:
        $ref: '#/definitions/entity.Group'
      group_id:
        type: integer
      hackerrank:
        type: string
      head_attendances:
        items:
          $ref: '#/definitions/entity.Attendance'
        type: array
      id:
        description: Core Identity Fields
        type: integer
      inactive:
        type: boolean
      instagram:
        type: string
      joined_date:
        type: string
      leetcode:
        description: |-
          Coding Profiles (optional)
          Use pointer types so that if no value is provided, they remain nil.
        type: string
      linkedin:
        description: Social Media (all optional)
        type: string
      mentor_name:
        type: string
      name:
        type: string
      phone:
        description: Contact Information
        type: string
      photo:
        description: System Fields
        type: string
      posts:
        items:
          $ref: '#/definitions/entity.Post'
        type: array
      preferred_language:
        type: string
      recent_actions:
        items:
          $ref: '#/definitions/entity.RecentAction'
        type: array
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        description: Role and Relationships
        type: integer
      short_bio:
        type: string
      student_id:
        type: string
      submissions:
        description: Relations (using GORM associations)
        items:
          $ref: '#/definitions/entity.Submission'
        type: array
      telegram_uid:
        type: string
      telegram_username:
        type: string
      tshirt_color:
        description: Physical Attributes
        type: string
      tshirt_size:
        type: string
      university:
        description: Academic Information
        type: string
      updated_at:
        type: string
    type: object
  entity.Vote:
    properties:
      comment:
        $ref: '#/definitions/entity.Comment'
      comment_id:
        type: integer
      contest:
        $ref: '#/definitions/entity.Contest'
      contest_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      post:
        $ref: '#/definitions/entity.Post'
      post_id:
        type: integer
      problem:
        $ref: '#/definitions/entity.Problem'
      problem_id:
        type: integer
      submission:
        $ref: '#/definitions/entity.Submission'
      submission_id:
        type: integer
      track:
        $ref: '#/definitions/entity.Track'
      track_id:
        type: integer
      type:
        description: e.g., 1=upvote, -1=downvote
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hub.a2sv.org
    name: API Support
  description: This is the API documentation for the Hub backend.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hub API
  version: "1.0"
paths:
...
  /api/super_to_groups:
    get:
      description: Get a list of all super to groups
      produces:
      - application/json
      responses:
        "200":
          description: List of super to groups
          schema:
            items:
              $ref: '#/definitions/entity.SuperToGroup'
            type: array
      summary: List super to groups
      tags:
      - SuperToGroups
    post:
      consumes:
      - application/json
      description: Create a new super to group entry
      parameters:
      - description: SuperToGroup data
        in: body
        name: super_to_group
        required: true
        schema:
          $ref: '#/definitions/entity.SuperToGroup'
      produces:
      - application/json
      responses:
        "201":
          description: Super to group created successfully
          schema:
            $ref: '#/definitions/entity.SuperToGroup'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new super to group
      tags:
      - SuperToGroups
  /api/super_to_groups/{id}:
    delete:
      description: Delete a super to group by its ID
      parameters:
      - description: SuperToGroup ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Super to group deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid super to group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Super to group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a super to group
      tags:
      - SuperToGroups
    get:
      description: Get a super to group by its ID
      parameters:
      - description: SuperToGroup ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Super to group details
          schema:
            $ref: '#/definitions/entity.SuperToGroup'
        "400":
          description: Invalid super to group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Super to group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get super to group by ID
      tags:
      - SuperToGroups
    patch:
      consumes:
      - application/json
      description: Update a super to group by its ID
      parameters:
      - description: SuperToGroup ID
        in: path
        name: id
        required: true
        type: integer
      - description: SuperToGroup data
        in: body
        name: super_to_group
        required: true
        schema:
          $ref: '#/definitions/entity.SuperToGroup'
      produces:
      - application/json
      responses:
        "200":
          description: Super to group updated successfully
          schema:
            $ref: '#/definitions/entity.SuperToGroup'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a super to group
      tags:
      - SuperToGroups
  /api/tracks:
    get:
      description: Get a list of all tracks
      produces:
      - application/json
      responses:
        "200":
          description: List of tracks
          schema:
            items:
              $ref: '#/definitions/entity.Track'
            type: array
      summary: List tracks
      tags:
      - Tracks
    post:
      consumes:
      - application/json
      description: Create a new track entry
      parameters:
      - description: Track data
        in: body
        name: track
        required: true
        schema:
          $ref: '#/definitions/entity.Track'
      produces:
      - application/json
      responses:
        "201":
          description: Track created successfully
          schema:
            $ref: '#/definitions/entity.Track'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new track
      tags:
      - Tracks
  /api/tracks/{id}:
    delete:
      description: Delete a track by its ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Track deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid track ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Track not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a track
      tags:
      - Tracks
    get:
      description: Get a track by its ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Track details
          schema:
            $ref: '#/definitions/entity.Track'
        "400":
          description: Invalid track ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Track not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get track by ID
      tags:
      - Tracks
    patch:
      consumes:
      - application/json
      description: Update a track by its ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: integer
      - description: Track data
        in: body
        name: track
        required: true
        schema:
          $ref: '#/definitions/entity.Track'
      produces:
      - application/json
      responses:
        "200":
          description: Track updated successfully
          schema:
            $ref: '#/definitions/entity.Track'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a track
      tags:
      - Tracks
  /api/tracks/name/{name}:
    get:
      description: Get a track by its name
      parameters:
      - description: Track Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Track details
          schema:
            $ref: '#/definitions/entity.Track'
        "400":
          description: Invalid track name
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Track not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get track by name
      tags:
      - Tracks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"